refactor: scope nodes.
