+++ src/utils/scopeNodes.ts
@@ -1,0 +1,12 @@
+const scopeNodes = [
+  'BlockStatement',
+  'FunctionDeclaration',
+  'FunctionExpression',
+  'ArrowFunctionExpression',
+  'ClassDeclaration',
+  'ClassExpression',
+  'ClassBody',
+  'StaticBlock',
+]
+
+export { scopeNodes }
